Added check when dividing by zero

+++ calculator_cli/src/main.rs
@@ -9,7 +9,7 @@ fn main() {
     io::stdin().read_line(&mut x)
     .expect("не получилось считать");
 
-    let x:i64 = match x.trim().parse() {
+    let x:f64 = match x.trim().parse() {
         Ok(num) => num,
         Err(_) => panic!("Введен символ, а не число"),
     };
@@ -19,7 +19,7 @@ fn main() {
     io::stdin().read_line(&mut y)
     .expect("не получилось считать");
 
-    let y:i64 = match y.trim().parse() {
+    let y:f64 = match y.trim().parse() {
         Ok(num) => num,
         Err(_) => panic!("Введен символ, а не число"),
     };
